% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmpraw}
\alias{jmpraw}
\title{JMP raw sanitation data}
\format{
A data frame with 16576 rows and 9 variables:
\describe{
\item{source}{Represents the code associated with the source of the data i.e census, report or survey}
\item{type}{Type of the data i.e census, survey or report}
\item{year}{Represents the year when the estimate was recorded}
\item{var_short}{Short variable sanitation description }
\item{value}{Represents the percentage pf the population using a particular sanitation service}
\item{iso3}{ISO3 country code}
\item{var_long}{Represents the sanitation service}
\item{residence}{Rural or urban residence}
\item{san_service_chain}{Represents the service ladder associated with the sanitation indicator}
}
}
\usage{
jmpraw
}
\description{
This dataset contains raw primary and secondary sanitation estimates.
}
\keyword{datasets}
